refactor: Update error.c, main.c, exec_cmd.c, free.c, and parse_args.c files
This commit updates multiple files in the codebase. The error.c file has been modified to include a new function parameter in the error_exit() function. The main.c file now includes a check for the number of command line arguments and calls the error_args() function if the condition is not met. The exec_cmd.c file has been updated to handle cases where the args array is empty, freeing memory and exiting with a failure code. The free.c file now includes additional functions to close and free pipes. The parse_args.c file has been modified to dynamically allocate memory for the cmds and pids arrays based on the number of commands passed as command line arguments.